+++ tubular/comparison.py
@@ -2,11 +2,16 @@ from __future__ import annotations
 
 import pandas as pd  # noqa: TCH002
 
-from tubular.base import BaseTwoColumnTransformer
-from tubular.mixins import BaseDropOriginalMixin
+from tubular.base import BaseTransformer
+from tubular.mixins import DropOriginalMixin, NewColumnNameMixin, TwoColumnMixin
 
 
-class EqualityChecker(BaseDropOriginalMixin, BaseTwoColumnTransformer):
+class EqualityChecker(
+    DropOriginalMixin,
+    NewColumnNameMixin,
+    TwoColumnMixin,
+    BaseTransformer,
+):
     """Transformer to check if two columns are equal.
 
     Parameters
@@ -14,7 +19,7 @@ class EqualityChecker(BaseDropOriginalMixin, BaseTwoColumnTransformer):
     columns: list
         List containing names of the two columns to check.
 
-    new_col_name: string
+    new_column_name: string
         string containing the name of the new column.
 
     drop_original: boolean = False
@@ -28,13 +33,15 @@ class EqualityChecker(BaseDropOriginalMixin, BaseTwoColumnTransformer):
     def __init__(
         self,
         columns: list,
-        new_col_name: str,
+        new_column_name: str,
         drop_original: bool = False,
         **kwargs: dict[str, bool],
     ) -> None:
-        super().__init__(columns=columns, new_col_name=new_col_name, **kwargs)
+        super().__init__(columns=columns, **kwargs)
 
-        BaseDropOriginalMixin.set_drop_original_column(self, drop_original)
+        self.check_two_columns(columns)
+        self.set_drop_original_column(drop_original)
+        self.check_and_set_new_column_name(new_column_name)
 
     def transform(self, X: pd.DataFrame) -> pd.DataFrame:
         """Create a column which is populated by the boolean
@@ -53,10 +60,10 @@ class EqualityChecker(BaseDropOriginalMixin, BaseTwoColumnTransformer):
         """
         X = super().transform(X)
 
-        X[self.new_col_name] = X[self.columns[0]] == X[self.columns[1]]
+        X[self.new_column_name] = X[self.columns[0]] == X[self.columns[1]]
 
         # Drop original columns if self.drop_original is True
-        BaseDropOriginalMixin.drop_original_column(
+        DropOriginalMixin.drop_original_column(
             self,
             X,
             self.drop_original,